refactor(ethereum_clis): rename python package with EthereumCLI
